Refactor deriva-csv code, tests, and dependencies (goodtables, et al) as an "extra" installable component.
