+++ tests/deriva/utils/catalog/manage/test_derivaCatalogToString.py
@@ -2,11 +2,10 @@ from unittest import TestCase
 import tempfile
 import sys
 import deriva.core.ermrest_model as em
-from deriva.utils.catalog.manage.utils import TempErmrestCatalog
+from deriva.utils.catalog.manage.utils import TempErmrestCatalog, load_module_from_path
 from deriva.core import get_credential
 from deriva.core.ermrest_config import tag as chaise_tags
 from deriva.utils.catalog.manage.dump_catalog import DerivaCatalogToString
-from deriva.utils.catalog.manage.deriva_csv import load_module_from_path
 
 from urllib.parse import urlparse
